# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Homer Next-Gen API
  version: "4.0.2-draft"
  description: >
    Draft Next-Gen REST API specification for the new homer-core.
servers:
  - url: "{apiPrefix}/api/v4"
    variables:
      apiPrefix:
        default: ""
        description: Optional API prefix (for example, /homer)
...
security:
  - bearerAuth: []
paths:
...
  /scripts:
    get:
      tags: [Settings]
      summary: List HEP scripts
      operationId: scriptsList
      parameters:
        - $ref: "#/components/parameters/PageLimit"
        - name: "filter[profile]"
          in: query
          schema:
            type: string
        - name: "filter[type]"
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of HEP scripts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HepScriptListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
    post:
      tags: [Settings]
      summary: Create HEP script (admin)
      operationId: scriptsCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HepScriptUpsert"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/ServerError"
  /scripts/{scriptId}:
    get:
      tags: [Settings]
      summary: Get HEP script by ID
      operationId: scriptsGet
      parameters:
        - name: scriptId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: HEP script
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HepScriptResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
    put:
      tags: [Settings]
      summary: Update HEP script (admin)
      operationId: scriptsUpdate
      parameters:
        - name: scriptId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HepScriptUpsert"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
    delete:
      tags: [Settings]
      summary: Delete HEP script (admin)
      operationId: scriptsDelete
      parameters:
        - name: scriptId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /agents/subscriptions:
    get:
      tags: [Agents]
      summary: Agent subscriptions
      operationId: agentSubscriptionsList
      parameters:
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageCursor"
        - $ref: "#/components/parameters/Sort"
        - $ref: "#/components/parameters/FilterAgentType"
        - $ref: "#/components/parameters/FilterAgentNode"
      responses:
        "200":
          description: Subscriptions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentSubListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
    post:
      tags: [Agents]
      summary: Create agent subscription (Auth-Token)
      operationId: agentSubscriptionsCreate
      security:
        - authTokenHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentSubUpsert"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
  /agents/subscriptions/{subscriptionId}:
    get:
      tags: [Agents]
      summary: Agent subscription by ID
      operationId: agentSubscriptionsGet
      parameters:
        - $ref: "#/components/parameters/AgentSubId"
      responses:
        "200":
          description: Subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentSubResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
    put:
      tags: [Agents]
      summary: Update agent subscription
      operationId: agentSubscriptionsUpdate
      parameters:
        - $ref: "#/components/parameters/AgentSubId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentSubUpsert"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
    delete:
      tags: [Agents]
      summary: Delete agent subscription
      operationId: agentSubscriptionsDelete
      parameters:
        - $ref: "#/components/parameters/AgentSubId"
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    authTokenHeader:
      type: apiKey
      in: header
      name: Auth-Token
  parameters:
...
    AgentSubId:
      name: subscriptionId
      in: path
      required: true
      schema:
        type: string
...
    PageLimit:
      name: page[limit]
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 1000
    PageCursor:
      name: page[cursor]
      in: query
      required: false
      schema:
        type: string
    Sort:
      name: sort
      in: query
      required: false
      description: "Sort, format: field or -field"
      schema:
        type: string
...
    FilterAgentType:
      name: filter[type]
      in: query
      required: false
      schema:
        type: string
    FilterAgentNode:
      name: filter[node]
      in: query
      required: false
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServerError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"
    Error:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          example: https://api.homer/errors/invalid_argument
        title:
          type: string
          example: Invalid argument
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: Field `from` must be <= `to`
        errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
    FieldError:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    Meta:
      type: object
      properties:
        request_id:
          type: string
        trace_id:
          type: string
        pagination:
          $ref: "#/components/schemas/Pagination"
        message:
          type: string
    Pagination:
      type: object
      properties:
        limit:
          type: integer
        next_cursor:
          type: string
        prev_cursor:
          type: string
        total:
          type: integer
        has_more:
          type: boolean
    IdResponse:
      type: object
      properties:
        data:
          type: string
        meta:
          $ref: "#/components/schemas/Meta"
...
    AgentSub:
      type: object
      properties:
        uuid:
          type: string
        gid:
          type: integer
        host:
          type: string
        port:
          type: integer
        protocol:
          type: string
        path:
          type: string
        node:
          type: string
        type:
          type: string
        ttl:
          type: integer
        create_date:
          type: string
        expire_date:
          type: string
        active:
          type: integer
    AgentSubUpsert:
      $ref: "#/components/schemas/AgentSub"
    AgentSubResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/AgentSub"
        meta:
          $ref: "#/components/schemas/Meta"
    AgentSubListResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/AgentSub"
        meta:
          $ref: "#/components/schemas/Meta"
...
    HepScriptUpsert:
      type: object
      properties:
        guid:
          type: string
        profile:
          type: string
        hep_alias:
          type: string
        type:
          type: string
        hepid:
          type: integer
        status:
          type: boolean
        script:
          type: string
    HepScript:
      type: object
      properties:
        id:
          type: integer
        guid:
          type: string
        profile:
          type: string
        hep_alias:
          type: string
        type:
          type: string
        hepid:
          type: integer
        status:
          type: boolean
        script:
          type: string
    HepScriptListResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/HepScript"
        meta:
          $ref: "#/components/schemas/Meta"
    HepScriptResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/HepScript"
        meta:
          $ref: "#/components/schemas/Meta"